split lookupPTSlot into different arch
