+++ sel4_vspace/src/arch/aarch64/structures.rs
@@ -1,4 +1,5 @@
-use crate::vm_attributes_t;
+use crate::{pte_t, vm_attributes_t};
+use sel4_common::structures::exception_t;
 
 impl vm_attributes_t {
     pub fn get_armExecuteNever(&self) -> bool {
@@ -16,4 +17,14 @@ impl vm_attributes_t {
             false
         }
     }
+}
+
+///lookup_pt_slot函数的返回值，
+/// `ptSlot`：找到的虚地址对应的`pte`的存放槽
+/// `ptBitsLeft`:找到叶子节点时，虚地址剩余未被索引的位置
+#[repr(C)]
+#[derive(Copy, Clone)]
+pub struct lookupPTSlot_ret_t {
+    pub status: exception_t,
+    pub ptSlot: *mut pte_t,
 }
