+++ sel4_vspace/src/arch/riscv64/structures.rs
@@ -1,0 +1,11 @@
+use crate::pte_t;
+
+///lookup_pt_slot函数的返回值，
+/// `ptSlot`：找到的虚地址对应的`pte`的存放槽
+/// `ptBitsLeft`:找到叶子节点时，虚地址剩余未被索引的位置
+#[repr(C)]
+#[derive(Copy, Clone)]
+pub struct lookupPTSlot_ret_t {
+    pub ptSlot: *mut pte_t,
+    pub ptBitsLeft: usize,
+}
